+++ wgpu/src/macros/mod.rs
@@ -60,7 +60,7 @@ macro_rules! _vertex_attr_array_helper {
                 offset: $off,
                 shader_location: $location,
             },];
-            $off + $crate::VertexFormat :: $format.size() as u32;
+            $off + $crate::VertexFormat :: $format.size();
             $($ll => $ii ,)*
         )
     };
@@ -73,7 +73,7 @@ fn test_vertex_attr_array() {
     assert_eq!(attrs.len(), 2);
     assert_eq!(attrs[0].offset, 0);
     assert_eq!(attrs[0].shader_location, 0);
-    assert_eq!(attrs[1].offset, size_of::<(f32, f32)>() as u32);
+    assert_eq!(attrs[1].offset, size_of::<(f32, f32)>() as u64);
     assert_eq!(attrs[1].shader_location, 3);
 }
